only delete old submissions 20% of the times.
the query that deletes past submissions (over the ARCHIVE limit)
is currently running on all form submissions for all forms.

this PR introduces basic randomness so it runs only on 20% of the
submissions.

if the form has 1000 stored submissions that query is very heavy,
if the form has less than 1000 it is totally unnecessary (while
still being possibly heavy).